handshake: revert addition of big-endian handshake format
Deprecating and eventually removing the current little-endian handshake
format just causes unnecessary churn for users without objective benefit.

The only field that was unconditionally encoded as big-endian before is
"error detail"; as this is also variable-length and was never longer
than 1 byte in practice (and it will not be for a long time), we can also
switch that around to little-endian without any compatiblity issues.