receive: always allocate a reasonably sized receive buffer
With a very small MTU and a great number of configured methods, the
receive buffer may be too small for a handshake packet. Always allocate
receive buffers of at least 1232 bytes, even with smaller MTU.

We also check that we never allocate a buffer for a bigger handshakes
and abort otherwise. This could be turned into a soft error in the
future, but so far we have a nice marging until we hit that number
(maximum handshake with all methods is ~640 bytes).