cipher: xmm: enable IBT and SHSTK support
Add endbr instructions for IBT (Indirect Branch Tracking). This is not
strictly necessary, as these functions are never called indirectly, but
there's also no reason not to. While fastd now requires a Binutils
version that supports these instructions, this is the case at least
since 2.30, which is older than fastd's minimum required Meson version.

For SHSTK (Shadow Stacks) no changes are needed.

A note section is inserted to mark both features as supported.